Messily solve day 17
This comes down to detecting a cycle and then not repeating that work.
We only have to do the head, plus enough cycles to recognize and count
the cycle (two), plus the tail.

This code is UGLY but it is after midnight and the idea is beautiful
(and that's what counts)