Day 21 part 2
This was super fun! I could have plugged it into WolframAlpha or
something, but it was actually quite simple to go down the tree with the
crucial assumption that "humn" was only called once (which must be true
for this linear equation to be solveable, since we only have one equation).

At each monkey, we have a desired value. One branch gives a constant and
the other contains the human. We have a very simple linear equation of
one variable which we can manually invert. Then we know what we want the
human-containing branch to equal, so we tail-recurse on that.