Day 16 part 2!
I read a blog post of someone else's solution, and their approach gave
me the insight I needed (aka I used their approach).

The crux of it was: build a map, where the key is the valves opened in
26 minutes, and the value is the best flow achieved with that
combination of valves. From that, you can determine which disjoint pair
of valve sets has the best total flow.

<https://nickymeuleman.netlify.app/garden/aoc2022-day16#part-2>